Fixes duplicate popups when confirming deleting proposals
Since we're using Turbolinks, we don't want to reload the JS from page
to page. It only needs to load on the initial page load. This fixes an
issue where one of the libraries was binding the confirmation popup
event to the delete button every time you went to a different page,
causing the popup to be displayed multiple times.