Refactor voting app to Ember.js
The initial server-side app had the data model I wanted, but I did not
like how it used one big form with a bunch of hacky view logic for the
actual voting process. This refactors it to an Ember.js app to make it
easier to add the voting features I wanted.